clean up, add tracking, add styles constants (#20)
* clean up, add tracking, add styles constants
* style editor copy polish
* dataset editor copy polish

+++ deploy.sh
@@ -1,5 +1,19 @@
 #!/bin/bash
-touch out/.nojekyll
-touch out/CNAME
-echo 'mayagao.com' >> out/CNAME
-git push origin `git subtree split --prefix out master`:gh-pages --force
+
+STATUS="$(git status)"
+
+if [[ $STATUS == *"nothing to commit, working directory clean"* ]]
+then
+    sed -i "" 'out/' ./.gitignore
+    touch out/.nojekyll
+    touch out/CNAME
+    echo 'mayagao.com' >> out/CNAME
+    git add .
+    git commit -m "Edit .gitignore to publish"
+    git push origin `git subtree split --prefix out master`:gh-pages --force
+    git reset --soft HEAD~1
+    git checkout .gitignore
+    echo "💎 Deploy success!"
+else
+    echo "Need clean working directory to deploy"
+fi
